Remove rocks/slk.hpp file
